ASKAPSDP-2230: Added remote ingest node services build. Now coordinates using semaphores to detect completion of parallel builds before deploying. Ready to test/debug.

+++ Tools/Dev/deploy/ASKAPSDP-deploy.sh
@@ -12,12 +12,10 @@ main() {
 
   doUserCheck
   
-  doLockCheck
-
   # Usage
   me=$(basename "$0")
   USAGE="$me [-h | ?] [-t <svn-tag>] [-l] [-c <cfg-file>]"
-
+  
   # Defaults, these can be overriden in config file (see -c switch)
   OVERWRITE_EXISTING=FALSE
   MODULES_DIR="/group/askap/modulefiles"
@@ -28,13 +26,13 @@ main() {
   DEPLOY_ASKAP_PIPLINE=true
   DEPLOY_ASKAP_SERVICES=true   # THIS NEEDS TO BE BUILT ON INGEST CLUSTER
   DEPLOY_ASKAP_CLI=true
-  NOW=`date "+%Y%m%d%H%M%S"`
-  NOW_LONG=`date`
+  MAKE_MODULES_DEFAULT=true
   REPO_URL="https://svn.atnf.csiro.au/"
   LOG_FILE="${NOW}-ASKAPSDP-deploy.log"
-  VERBOSE=false
 
   # Globals
+  NOW=`date "+%Y%m%d%H%M%S"`
+  NOW_LONG=`date`
   CFG_FILE="RELEASE.cfg"
   CURRENT_DIR=$PWD
   TAG=""
@@ -43,7 +41,9 @@ main() {
   ASKAPSOFT_URL="${REPO_URL}/askapsdp/"
   TOS_URL="${REPO_URL}/askapsoft/trunk/"
   BUILD_DIR=""
-
+  VERBOSE=false
+  INGEST_NODE=galaxy-ingest01.pawsey.org.au
+  
   # Errors
   NO_ERROR=0
   ERROR_BAD_OVERRIDES=64
@@ -53,9 +53,16 @@ main() {
   ERROR_MISSING_CFG_FILE=68
   ERROR_NOT_ASKAPOPS=69
   ERROR_ALREADY_RUNNING=70
+  ERROR_WRONG_CLUSTER=71
+  TOLD_TO_DIE=72
 
-  # Process command line params
+  trap "{ cleanup ${TOLD_TO_DIE}; }" SIGHUP
+
   processCmdLine "$@" 
+
+  doLockCheck
+  
+  detectHost
 
   # Overrides are in the cfg file
   if [ -e $CFG_FILE ]; then
@@ -98,6 +105,9 @@ main() {
   if [ "$DEPLOY_ASKAP_CLI" == true ]; then
     buildCli
   fi
+  
+  # Stage the build
+  stage
 
   # Deploy the modules
   deploy
@@ -123,40 +133,19 @@ buildAskap() {
   if [ "$DEPLOY_ASKAP_SERVICES" == true ]; then
     cp -R ../askapsoft-${TAG} ../askapingest-${TAG}
   fi
-
-  # Initialise rbuild env
-  python2.7 bootstrap.py -n
-  source initaskap.sh
-
-  # build askapsoft if specified; produces a deploy dir in build ready to copy
+  
+  # Create a build command which will run initialisation and all required builds in the background.
+  local BUILD_COMMAND="python2.7 bootstrap.py -n;
+                       source initaskap.sh; "
+  
+  # build askapsoft if specified
   if [ "$DEPLOY_ASKAPSOFT" == true ]; then
-    rbuild -n -V --release-name=ASKAPsoft-cpapps-${TAG} --stage-dir=ASKAPsoft-cpapps-${TAG} -t release ${ASKAP_ROOT}/Code/Systems/cpapps
-    tar xf ASKAPsoft-cpapps-${TAG}.tgz
-    cd ASKAPsoft-cpapps-${TAG}; rm -rf build-1 config docs ICE_LICENSE include initenv.py local man share slice ssl; cd ..
-    
-    # fix up permissions
-    chgrp -R askap ASKAPsoft-cpapps-${TAG}
-    find ASKAPsoft-cpapps-${TAG} -type f | xargs chmod oug+r
-    find ASKAPsoft-cpapps-${TAG} -type d | xargs chmod oug+rx
-    
-    # stage
-    mv ASKAPsoft-cpapps-${TAG} ${BUILD_DIR}
-    
+    BUILD_COMMAND="${BUILD_COMMAND} rbuild -n -V --release-name ASKAPsoft-cpapps-${TAG} --stage-dir ASKAPsoft-cpapps-${TAG} -t release Code/Systems/cpapps; "
   fi
 
   # build pipelines if specified
   if [ "$DEPLOY_ASKAP_PIPLINE" == true ]; then
-    rbuild -n -V --release-name ASKAPsoft-pipelines-${TAG} --stage-dir ASKAPsoft-pipelines-${TAG} -t release ${ASKAP_ROOT}/Code/Systems/pipelines
-    tar xf ASKAPsoft-pipelines-${TAG}.tgz
-    
-    # fix up permissions
-    chgrp -R askap ASKAPsoft-pipelines-${TAG}
-    find ASKAPsoft-pipelines-${TAG} -type f | xargs chmod oug+r
-    find ASKAPsoft-pipelines-${TAG} -type d | xargs chmod oug+rx
-    
-    # stage
-    mv ASKAPsoft-pipelines-${TAG} ${BUILD_DIR}
-    
+    BUILD_COMMAND="${BUILD_COMMAND} rbuild -n -V --release-name ASKAPsoft-pipelines-${TAG} --stage-dir ASKAPsoft-pipelines-${TAG} -t release Code/Systems/pipelines; "
   fi
 
   # build askapdoc if specified
@@ -169,11 +158,13 @@ buildAskap() {
       svn ci -m "Updated version numbers for " Code/Doc/user/current/doc/conf.py
     fi
     
-    rbuild -n -t doc Code/Doc/user/current
-    mv Code/Doc/user/current/doc/_build/html ${BUILD_DIR}
+    BUILD_COMMAND="${BUILD_COMMAND} rbuild -n -t doc Code/Doc/user/current; "
   fi
   
-  #rbuild -t doc ${ASKAP_ROOT}/Code/Doc/user/current
+  BUILD_COMMAND="${BUILD_COMMAND} touch ${BUILD_DIR}/build-askapsoft-done; "
+  
+  (eval $BUILD_COMMAND) &
+
 }
 
 # Build the ASKAP services including the CP Manager. This will need to be built on ingest cluster.
@@ -182,6 +173,8 @@ buildAskap() {
 buildServices() {
   log "Building the ASKAP services\n"
 
+  cd ${WORKING_DIR}/${TAG}
+
   # Check it out of it's not there
   if [ ! -e askapingest-${TAG} ]; then
     svn export -q ${ASKAPSOFT_URL} askapingest-${TAG}
@@ -189,7 +182,14 @@ buildServices() {
 
   # This needs to be done on the ingest nodes
   # build ingest if specified
-  #rbuild -n -V -t release ${ASKAP_ROOT}/Code/Systems/cpsvcs
+  local BUILD_COMMAND="(cd ${WORKING_DIR}/${TAG}/askapingest-${TAG};
+                        python2.7 bootstrap.py -n;
+                        source initaskap.sh; 
+                        rbuild -n -V --release-name ASKAPsoft-cpsvcs-${TAG} --stage-dir ASKAPsoft-cpsvcs-${TAG} -t release Code/Systems/cpsvcs;
+                        touch ${BUILD_DIR}/build-cpsvcs-done) "
+                       
+  # run this on ingest
+  ssh -v -t askapops@${INGEST_NODE} "${BUILD_COMMAND}"
 }
 
 # Build the CLI tools. A checkout of the TOS trunk is required for this.
@@ -198,28 +198,124 @@ buildServices() {
 buildCli() {
   log "Building the CLI tools\n"
 
-  # build askapcli if specified
-  if [ "$DEPLOY_ASKAP_CLI" == true ]; then
-    svn export -q ${TOS_URL} askaptos-trunk
-    cd askaptos-trunk
-    
-    # Initialise rbuild env
-    python2.7 bootstrap.py -n
-    source initaskap.sh
+  cd ${WORKING_DIR}/${TAG}
 
-    rbuild -n -V --release-name=ASKAPsoft-cli-${TAG} --stage-dir=ASKAPsoft-cli-${TAG} -t release ${ASKAP_ROOT}/Code/Components/UI/cli/current
-    tar xf ASKAPsoft-cli-${TAG}
+  # Create a build command which will run initialisation and all required builds in the background.
+  local BUILD_COMMAND="python2.7 bootstrap.py -n;
+                       source initaskap.sh; 
+                       rbuild -n -V --release-name ASKAP-cli-${TAG} --stage-dir ASKAP-cli-${TAG} -t release Code/Components/UI/cli/current; 
+                       touch ${BUILD_DIR}/build-cli-done"
+                       
+  svn export -q ${TOS_URL} askaptos-trunk
+  cd askaptos-trunk
+    
+  (eval $BUILD_COMMAND) &
+
+}
+
+# Wait for the artefacts to be built and then stage the build
+stage() {
+  cd ${BUILD_DIR}
+
+  # Wait for builds to finish
+  # true if the build-xxxx-done files are present for the required builds (including the services one running on ingest)
+  local ALL_DONE=false
+  local WAIT_TIME=60
+  local declare -a SEMAPHORES
+  local IDX=0
+  if [ "$DEPLOY_ASKAPSOFT" == true || "$DEPLOY_ASKAP_PIPLINE" == true || "$DEPLOY_ASKAP_USER_DOC" == true ]; then
+    SEMAPHORES[$IDX]="build-askapsoft-done"
+    IDX=IDX+1
+  fi
+  if [ "$DEPLOY_ASKAP_CLI" == true ]; then
+    SEMAPHORES[$IDX]="build-cli-done"
+    IDX=IDX+1
+  fi
+  if [ "$DEPLOY_ASKAP_SERVICES" == true ]; then
+    SEMAPHORES[$IDX]="build-cpsvcs-done"
+    IDX=IDX+1
+  fi
+  
+  until [ $ALL_DONE == true ]; do
+  
+    # check all the semaphores
+    for FLAG in ${SEMAPHORES[@]}; do
+      if [ -e $FLAG ]; then
+        ALL_DONE=true
+      else
+        ALL_DONE=false
+      fi
+    done
+    
+    # if any were not done then wait again
+    if [ $ALL_DONE == false ]; then
+      sleep $WAIT_TIME
+    fi
+  done
+
+  # stage askapsoft if specified
+  if [ "$DEPLOY_ASKAPSOFT" == true ]; then
+    tar xf askapsoft-${TAG}/ASKAPsoft-cpapps-${TAG}.tgz
+    cd ASKAPsoft-cpapps-${TAG}; rm -rf build-1 config docs ICE_LICENSE include initenv.py local man share slice ssl; cd ..
     
     # fix up permissions
-    chgrp -R askap ASKAPsoft-cli-${TAG}
-    find ASKAPsoft-cli-${TAG} -type f | xargs chmod oug+r
-    find ASKAPsoft-cli-${TAG} -type d | xargs chmod oug+rx
+    chgrp -R askap ASKAPsoft-cpapps-${TAG}
+    find ASKAPsoft-cpapps-${TAG} -type f | xargs chmod oug+r
+    find ASKAPsoft-cpapps-${TAG} -type d | xargs chmod oug+rx
     
     # stage
-    mv ASKAPsoft-cli-${TAG} ${BUILD_DIR}
+    mv ASKAPsoft-cpapps-${TAG} ${BUILD_DIR}
+    sed -e "s/\${_VERSION}.*/${VERSION}/" askapsoft-${TAG}/Tools/Dev/deploy/askapsoft.module-template > ${BUILD_DIR}/askapsoft.module
+    
+  fi
+
+  # stage pipelines if specified
+  if [ "$DEPLOY_ASKAP_PIPLINE" == true ]; then
+    tar xf ${askapsoft-${TAG}/ASKAPsoft-pipelines-${TAG}.tgz
+    
+    # fix up permissions
+    chgrp -R askap ASKAPsoft-pipelines-${TAG}
+    find ASKAPsoft-pipelines-${TAG} -type f | xargs chmod oug+r
+    find ASKAPsoft-pipelines-${TAG} -type d | xargs chmod oug+rx
+    
+    # stage
+    mv ASKAPsoft-pipelines-${TAG} ${BUILD_DIR}
+    sed -e "s/\${_VERSION}.*/${VERSION}/" askapsoft-${TAG}/Tools/Dev/deploy/askappipeline.module-template > ${BUILD_DIR}/askappipeline.module
 
   fi
 
+  # stage askapcli if specified
+  if [ "$DEPLOY_ASKAP_CLI" == true ]; then
+    tar xf askaptos-trunk/ASKAP-cli-${TAG}
+
+    # fix up permissions
+    chgrp -R askap ASKAP-cli-${TAG}
+    find ASKAP-cli-${TAG} -type f | xargs chmod oug+r
+    find ASKAP-cli-${TAG} -type d | xargs chmod oug+rx
+    
+    mv ASKAP-cli-${TAG} ${BUILD_DIR}
+    sed -e "s/\${_VERSION}.*/${VERSION}/" askapsoft-${TAG}/Tools/Dev/deploy/askapcli.module-template > ${BUILD_DIR}/askapcli.module
+  fi
+
+  # stage askapdoc if specified
+  if [ "$DEPLOY_ASKAP_USER_DOC" == true ]; then
+    mv askapsoft-${TAG}/Code/Doc/user/current/doc/_build/html ${BUILD_DIR}
+  fi
+  
+  # stage askapservices if specified
+  if [ "$DEPLOY_ASKAP_SERVICES" == true ]; then
+    tar xf askapingest-${TAG}/ASKAPsoft-cpapps-${TAG}.tgz
+    cd ASKAPsoft-cpapps-${TAG}; rm -rf build-1 config docs ICE_LICENSE include initenv.py local man share slice ssl; cd ..
+    
+    # fix up permissions
+    chgrp -R askap ASKAPsoft-cpapps-${TAG}
+    find ASKAPsoft-cpapps-${TAG} -type f | xargs chmod oug+r
+    find ASKAPsoft-cpapps-${TAG} -type d | xargs chmod oug+rx
+    
+    # stage
+    mv ASKAPsoft-cpapps-${TAG} ${BUILD_DIR}
+    sed -e "s/\${_VERSION}.*/${VERSION}/" askapingest-${TAG}/Tools/Dev/deploy/askapservices.module-template > ${BUILD_DIR}/askapservices.module
+  fi
 }
 
 # Deploy the built modules; some might not have been built, but the ones that have are in the staging dir (build).
@@ -227,23 +323,40 @@ buildCli() {
 deploy() {
   log "Deploying the ASKAP modules to ${DEPLOY_DIR}\n"
 
-  # Copy the askap apps if they were built
+  # Deploy the askap apps, if they were built, along with module configuration file
   if [ -e ${BUILD_DIR}/ASKAPsoft-cpapps-${TAG} ]; then
     mv ${BUILD_DIR}/ASKAPsoft-cpapps-${TAG} ${DEPLOY_DIR}/askapsoft/${TAG}
+    mv ${BUILD_DIR}/askapsoft.module ${MODULES_DIR}/askapsoft/${VERSION}
+    if [ "$MAKE_DEFAULT" == true ]; then
+      # change the default referenced version
+      sed -i -e "s/^set ModulesVersion.*/set ModulesVersion \"${VERSION}\"/" ${MODULES_DIR}/askapsoft/.version
+    fi
   fi
+  
+  # Deploy the askap pipeline scripts, if they were built, along with module configuration file
   if [ -e ${BUILD_DIR}/ASKAPsoft-pipelines-${TAG} ]; then
     mv ${BUILD_DIR}/ASKAPsoft-pipelines-${TAG} ${DEPLOY_DIR}/askappipeline/${TAG}
+    mv ${BUILD_DIR}/askappipeline.module ${MODULES_DIR}/askappipeline/${VERSION}
+    if [ "$MAKE_DEFAULT" == true ]; then
+      # change the default referenced version
+      sed -i -e "s/^set ModulesVersion.*/set ModulesVersion \"${VERSION}\"/" ${MODULES_DIR}/askappipeline/.version
+    fi
   fi
-  if [ -e ${BUILD_DIR}/ASKAPsoft-cli-${TAG} ]; then
-    mv ${BUILD_DIR}/ASKAPsoft-cli-${TAG} ${DEPLOY_DIR}/askapcli/${TAG}
+  
+  # Deploy the askap cli, if it was built, module configuration file is always 'current'
+  if [ -e ${BUILD_DIR}/ASKAP-cli-${TAG} ]; then
+    mv ${BUILD_DIR}/ASKAP-cli-${TAG} ${DEPLOY_DIR}/askapcli/${TAG}
     rm ${DEPLOY_DIR}/askapcli/current
     ln -s ${DEPLOY_DIR}/askapcli/${TAG} ${DEPLOY_DIR}/askapcli/current
   fi
 }
 
-# Cleanup
+# Cleanup. Only call this once we are definitely running, so after command line processing and lock check.
 cleanup () {
-  rm /tmp/ASKAPDEPLOY.lock
+  if [ -e /tmp/ASKAPDEPLOY.lock ]; then
+    rm /tmp/ASKAPDEPLOY.lock
+    #rm -rf ${WORKING_DIR}/${TAG}
+  fi
   exit $1
 }
 
@@ -251,10 +364,26 @@ showHelp() {
   printf "$USAGE \n"
   printf "Make a release of ASKAPsoft.\n"
   printf "    -h | ?            this help\n"
-  printf "    -v                verbose"
+  printf "    -v                verbose\n"
   printf "    -t svn-tag        is optional, if not given must specify -l\n"
   printf "    -l                use the latest tag in the repository (caution)\n"
   printf "    -c cfg-file       is optional, if not given and a file RELEASE.cfg is not present in the current dir then no overrides will be used\n"
+  printf "\n"
+  printf "You can override the following settings by editing a scriptable config file and passing it via -c switch \n"
+  printf "(or use the default RELEASE.cfg). This script will be sourced to set up the deploy environment. \n"
+  printf "Unless overriden in this way, the defaults are:\n\n"
+  printf "  export WORKING_DIR='/group/astronomy856/askapops/workspace/askapsdp/'\n"
+  printf "  export OVERWRITE_EXISTING=false\n"
+  printf "  export DEPLOY_DIR='/group/askap/'\n"
+  printf "  export MODULES_DIR='/group/askap/modulefiles'\n"
+  printf "  export DEPLOY_ASKAPSOFT=true\n"
+  printf "  export DEPLOY_ASKAP_USER_DOC=true\n"
+  printf "  export DEPLOY_ASKAP_PIPLINE=true\n"
+  printf "  export DEPLOY_ASKAP_SERVICES=true\n"
+  printf "  export DEPLOY_ASKAP_CLI=true\n"
+  printf "  export MAKE_MODULES_DEFAULT=true\n"
+  printf "  export REPO_URL='https://svn.atnf.csiro.au/'\n"
+  printf "  export LOG_FILE='./${NOW}-ASKAPSDP-deploy.log'\n"
 }
 
 # Logger for logging messages
@@ -288,9 +417,10 @@ processCmdLine() {
   # Process our options
   while getopts "h?vlt:c:" opt "$@"; do
       case "$opt" in
+      
       h|\?)
           showHelp
-          cleanup $NO_ERROR
+          exit $NO_ERROR
           ;;
       v)
           VERBOSE=true
@@ -300,7 +430,7 @@ processCmdLine() {
             TAG=$OPTARG
           else
             printf "The tag %s does not match semantic versioning string format [CP-0.12.3]\n" ${OPTARG}
-            cleanup $ERROR_BAD_SVN_TAG;
+            exit $ERROR_BAD_SVN_TAG;
           fi
           ;;
       l)
@@ -311,7 +441,7 @@ processCmdLine() {
             CFG_FILE=$OPTARG
           else
             printf "The given config file %s does not exist\n" ${OPTARG}
-            cleanup $ERROR_MISSING_CFG_FILE;
+            exit $ERROR_MISSING_CFG_FILE;
           fi
           ;;
       esac
@@ -324,7 +454,7 @@ processCmdLine() {
   if [ "$TAG" == "" ]; then 
     printf "No TAG has been specified\n"
     printf "${USAGE}\n"
-    cleanup $ERROR_NO_TAG;
+    exit $ERROR_NO_TAG;
   fi
 
 }
@@ -349,6 +479,7 @@ findLatestTag() {
 # Print the environment to the log file
 logENV () {
   log "OVERWRITE_EXISTING = %s\n" $([ "$OVERWRITE_EXISTING" == true ] && echo "true" || echo "false")
+  log "MAKE_MODULES_DEFAULT = %s\n" $([ "$MAKE_MODULES_DEFAULT" == true ] && echo "true" || echo "false")
   log "MODULES_DIR = %s\n" $MODULES_DIR
   log "WORKING_DIR = %s\n" $WORKING_DIR
   log "DEPLOY_DIR = %s\n" $DEPLOY_DIR
@@ -414,20 +545,33 @@ doOverwriteChecks() {
   fi
 }
 
+# We only want this to run as the asksapops user
 doUserCheck () {
-  # display usage if the script is not run as askapops user 
   if [[ $USER != "bas091" ]]; then 
     logerr "This script must be run as askapops!" 
     exit $ERROR_NOT_ASKAPOPS;
   fi 
 }
 
+# Handling a semaphore file to prevent running build twice
 doLockCheck () {
   if [ -e /tmp/ASKAPDEPLOY.lock ]; then
-    logerr "This script is already running - aborting!" 
+    printf "ERROR: This script is already running - aborting!\n"  1>&2
     exit $ERROR_ALREADY_RUNNING;
   fi
   touch /tmp/ASKAPDEPLOY.lock
+}
+
+# We need to know if we are running on galaxy or ingest cluster
+detectHost () {
+  shopt -s extglob
+
+  if [[ ${HOSTNAME} =~ "galaxy-??" ]]; then
+    GALAXY_CLUSTER=true;
+  else
+     printf "ERROR: Cannot run on this machine, must be galaxy or ingest cluster!\n"
+     exit $ERROR_WRONG_CLUSTER;
+  fi
 }
 
 main "$@"
